v1.1: implemented local logging and graphs (and bugfixes)
Meter data will now be logged on the module's LittleFS file system in hourly, daily (for a week, month and entire year) format. Logfiles can be downloaded and displayed in graphical form locally, so no need to hand over your meter data to a third party or run HA, Domoticz or similar to get an overview of your energy usage.